feat: remove old `ResponseDecoder`
